% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{not_all_na}
\alias{not_all_na}
\title{Check whether a vector contains at least one non-NA value}
\usage{
not_all_na(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Logical scalar.
}
\description{
Useful as a column selection predicate:
\code{df \%>\% select(where(not_all_na))}.
}
\examples{
not_all_na(c(NA, NA, 1))
# [1] TRUE
not_all_na(c(NA, NA))
# [1] FALSE
}
